pyproject.toml
[project]
name = "smart-life-planner"
version = "0.1.0"
description = "Smart Life Planner - Multi-agent planning system"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.28.0",
    "pydantic>=2.0.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
